problem-solver-assistant/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
openai==1.77.0
python-dotenv==1.0.1
pydantic==2.6.3 